# pkg/docs/swagger.yaml
basePath: /
definitions:
  models.Client:
    properties:
      birthday:
        type: string
      name:
        type: string
      sex:
        type: string
      uid:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This API provides endpoints to manage client data including loading
    data from a CSV file into PostgreSQL, and retrieving client information by UID,
    filtering by age range, and searching by name. The service is built with Golang
    using the Gin framework and provides Swagger documentation for easy API exploration.
  title: Welcome to the Jungle - Client API
  version: "1.0"
paths:
  /info:
    get:
      description: Get client info by UID
      parameters:
      - description: Client UID
        in: query
        name: uid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get client info
  /info/by-age:
    get:
      description: Get clients by age range
      parameters:
      - description: Start Date
        format: date
        in: query
        name: start
        required: true
        type: string
      - description: End Date
        format: date
        in: query
        name: end
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Client'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get clients by age
  /search:
    get:
      description: Search clients by name
      parameters:
      - description: Name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Client'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search clients by name
swagger: "2.0"